MediMarc/backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
python-decouple
django-filter
django-cors-headers
setuptools
resend
reportlab
